# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/expenses/split-preview": {
      "post": {
        "operationId": "ExpensesController_preview",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Authoritative split preview.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SplitPreviewEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Preview an authoritative deterministic split without posting it",
        "tags": ["expenses"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpenseMutationRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/expenses": {
      "post": {
        "operationId": "ExpensesController_create",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "Stable 8–200 character key. Reuse is valid only with the identical request body.",
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 200
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Expense posted.",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                }
              },
              "Idempotency-Replayed": {
                "schema": {
                  "type": "string",
                  "enum": ["true"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "summary": "Post an expense, journal, projections, outbox, and idempotency outcome atomically",
        "tags": ["expenses"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpenseMutationRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/expenses/{expenseId}": {
      "put": {
        "operationId": "ExpensesController_update",
        "parameters": [
          {
            "name": "expenseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{8}(?:-[0-9a-fA-F]{4}){3}-[0-9a-fA-F]{12}$",
              "example": "2f40d889-89f7-4fcb-99f0-8702f88d0c77"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "Stable 8–200 character key. Reuse is valid only with the identical request body.",
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 200
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "description": "Current positive decimal expense version, optionally enclosed in double quotes.",
            "schema": {
              "type": "string",
              "pattern": "^(?:\"[1-9][0-9]{0,18}\"|[1-9][0-9]{0,18})$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Expense replaced by an immutable revision and balanced journal reversal.",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "Idempotency-Replayed": {
                "schema": {
                  "type": "string",
                  "enum": ["true"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "summary": "Replace an expense by appending a revision and reversing its prior journal",
        "tags": ["expenses"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpenseMutationRequest"
              }
            }
          }
        }
      },
      "get": {
        "operationId": "ExpensesController_detail",
        "parameters": [
          {
            "name": "expenseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{8}(?:-[0-9a-fA-F]{4}){3}-[0-9a-fA-F]{12}$",
              "example": "2f40d889-89f7-4fcb-99f0-8702f88d0c77"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Expense details.",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Get an authorized expense with its current revision, payers, and shares",
        "tags": ["expenses"]
      }
    },
    "/api/v1/groups/{groupId}/expenses": {
      "get": {
        "operationId": "GroupExpensesController_list",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{8}(?:-[0-9a-fA-F]{4}){3}-[0-9a-fA-F]{12}$",
              "example": "2f40d889-89f7-4fcb-99f0-8702f88d0c77"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 1000
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 30
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Expense page.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpensePageEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "List posted group expenses with stable cursor pagination",
        "tags": ["groups", "expenses"]
      }
    },
...
  },
  "info": {
    "title": "SPLITO API",
    "description": "Versioned API for SPLITO expense sharing and its financial journal",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Current SPLITO deployment"
    }
  ],
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "SPLITO_SESSION"
      },
      "csrf": {
        "type": "apiKey",
        "in": "header",
        "name": "x-csrf-token"
      }
    },
    "schemas": {
...
      "ErrorEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": ["code", "message", "requestId"],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string",
                "pattern": "^[0-9a-fA-F]{8}(?:-[0-9a-fA-F]{4}){3}-[0-9a-fA-F]{12}$",
                "example": "2f40d889-89f7-4fcb-99f0-8702f88d0c77"
              },
              "fieldErrors": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": ["field", "message"],
                  "properties": {
                    "field": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "SplitPreviewLine": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "participantId",
          "displayName",
          "paidAmountMinor",
          "owedAmountMinor",
          "netAmountMinor"
        ],
        "properties": {
          "participantId": {
            "type": "string",
            "pattern": "^[0-9a-fA-F]{8}(?:-[0-9a-fA-F]{4}){3}-[0-9a-fA-F]{12}$",
            "example": "2f40d889-89f7-4fcb-99f0-8702f88d0c77"
          },
          "displayName": {
            "type": "string"
          },
          "paidAmountMinor": {
            "type": "string",
            "pattern": "^(?:0|[1-9][0-9]{0,18})$",
            "description": "Integer minor units serialized as a decimal string.",
            "example": "1250"
          },
          "owedAmountMinor": {
            "type": "string",
            "pattern": "^(?:0|[1-9][0-9]{0,18})$",
            "description": "Integer minor units serialized as a decimal string.",
            "example": "1250"
          },
          "netAmountMinor": {
            "type": "string",
            "pattern": "^(?:0|-?[1-9][0-9]{0,18})$",
            "description": "Signed integer minor units serialized as a decimal string.",
            "example": "-1250"
          }
        }
      },
      "SplitPreview": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "currency",
          "totalAmountMinor",
          "allocations",
          "explanation",
          "algorithmVersion"
        ],
        "properties": {
          "currency": {
            "type": "string",
            "pattern": "^[A-Z]{3}$",
            "example": "INR"
          },
          "totalAmountMinor": {
            "type": "string",
            "pattern": "^(?:0|[1-9][0-9]{0,18})$",
            "description": "Integer minor units serialized as a decimal string.",
            "example": "1250"
          },
          "allocations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SplitPreviewLine"
            }
          },
          "explanation": {
            "type": "string"
          },
          "algorithmVersion": {
            "type": "string"
          }
        }
      },
      "ExpensePayer": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "displayName", "paidAmountMinor"],
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[0-9a-fA-F]{8}(?:-[0-9a-fA-F]{4}){3}-[0-9a-fA-F]{12}$",
            "example": "2f40d889-89f7-4fcb-99f0-8702f88d0c77"
          },
          "displayName": {
            "type": "string"
          },
          "paidAmountMinor": {
            "type": "string",
            "pattern": "^(?:0|[1-9][0-9]{0,18})$",
            "description": "Integer minor units serialized as a decimal string.",
            "example": "1250"
          }
        }
      },
      "ExpenseAllocation": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "displayName", "owedAmountMinor", "netAmountMinor"],
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[0-9a-fA-F]{8}(?:-[0-9a-fA-F]{4}){3}-[0-9a-fA-F]{12}$",
            "example": "2f40d889-89f7-4fcb-99f0-8702f88d0c77"
          },
          "displayName": {
            "type": "string"
          },
          "owedAmountMinor": {
            "type": "string",
            "pattern": "^(?:0|[1-9][0-9]{0,18})$",
            "description": "Integer minor units serialized as a decimal string.",
            "example": "1250"
          },
          "netAmountMinor": {
            "type": "string",
            "pattern": "^(?:0|-?[1-9][0-9]{0,18})$",
            "description": "Signed integer minor units serialized as a decimal string.",
            "example": "-1250"
          },
          "inputValue": {
            "type": "string"
          }
        }
      },
      "Expense": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "description",
          "amount",
          "expenseDate",
          "category",
          "groupId",
          "groupName",
          "status",
          "version",
          "payers",
          "allocations",
          "revisionNumber",
          "createdBy",
          "canEdit"
        ],
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[0-9a-fA-F]{8}(?:-[0-9a-fA-F]{4}){3}-[0-9a-fA-F]{12}$",
            "example": "2f40d889-89f7-4fcb-99f0-8702f88d0c77"
          },
          "description": {
            "type": "string"
          },
          "amount": {
            "type": "object",
            "additionalProperties": false,
            "required": ["amountMinor", "currency"],
            "properties": {
              "amountMinor": {
                "type": "string",
                "pattern": "^(?:0|[1-9][0-9]{0,18})$",
                "description": "Integer minor units serialized as a decimal string.",
                "example": "1250"
              },
              "currency": {
                "type": "string",
                "pattern": "^[A-Z]{3}$",
                "example": "INR"
              }
            }
          },
          "expenseDate": {
            "type": "string",
            "format": "date",
            "example": "2026-09-12"
          },
          "category": {
            "type": "string"
          },
          "groupId": {
            "type": "string",
            "pattern": "^[0-9a-fA-F]{8}(?:-[0-9a-fA-F]{4}){3}-[0-9a-fA-F]{12}$",
            "example": "2f40d889-89f7-4fcb-99f0-8702f88d0c77"
          },
          "groupName": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["draft", "posted", "voided"]
          },
          "version": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "payers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExpensePayer"
            }
          },
          "allocations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExpenseAllocation"
            }
          },
          "revisionNumber": {
            "type": "integer",
            "minimum": 1
          },
          "splitMethod": {
            "type": "string",
            "enum": ["equal", "exact", "percentage", "shares", "adjustments"]
          },
          "algorithmVersion": {
            "type": "string"
          },
          "createdBy": {
            "type": "object",
            "additionalProperties": false,
            "required": ["id", "displayName"],
            "properties": {
              "id": {
                "type": "string",
                "pattern": "^[0-9a-fA-F]{8}(?:-[0-9a-fA-F]{4}){3}-[0-9a-fA-F]{12}$",
                "example": "2f40d889-89f7-4fcb-99f0-8702f88d0c77"
              },
              "displayName": {
                "type": "string"
              }
            }
          },
          "canEdit": {
            "type": "boolean",
            "description": "True only for the active member who originally created this posted expense."
          }
        }
      },
      "ExpensePage": {
        "type": "object",
        "additionalProperties": false,
        "required": ["items"],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Expense"
            }
          },
          "nextCursor": {
            "type": "string"
          }
        }
      },
...
      "SplitPreviewEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": ["data"],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SplitPreview"
          }
        }
      },
      "ExpenseEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": ["data"],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Expense"
          }
        }
      },
      "ExpensePageEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": ["data"],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ExpensePage"
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Malformed input or identifier.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "A valid session is required.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Forbidden": {
        "description": "The authenticated identity is not authorized.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "NotFound": {
        "description": "The resource does not exist or is not visible to this identity.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Conflict": {
        "description": "The resource version, idempotency body, or financial state conflicts.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "PreconditionFailed": {
        "description": "The supplied resource version is stale.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "PayloadTooLarge": {
        "description": "The uploaded file or multipart request exceeds its limit.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "UnsupportedMediaType": {
        "description": "The upload is not an accepted image or multipart body.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "UnprocessableEntity": {
        "description": "The financial-domain invariant was not satisfied.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "PreconditionRequired": {
        "description": "A required optimistic-concurrency header is missing.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "RateLimited": {
        "description": "The request rate limit was exceeded.",
        "headers": {
          "Retry-After": {
            "description": "Minimum seconds before retrying; 60 for cooldown or up to 900 for quota.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 900
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "InternalError": {
        "description": "An unexpected server error occurred.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "BadGateway": {
        "description": "The external SMS provider did not accept the invitation message.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "A required external delivery adapter is unavailable.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    }
  }
}